+++ src/app/components/Organism/AdminClassForm.tsx
@@ -5,7 +5,7 @@ import { db } from '@/firebase/config'
 import { addDoc, collection, setDoc, doc, updateDoc } from 'firebase/firestore'
 
 import Button from '@/app/components/Button'
-import ImageButton from '@/app/components/Molecule/ImageButton'
+import ImageButton from '@/app/components/Atom/Button/ImageButton'
 import Badge from '@/app/components/Molecule/Badge'
 
 import styled from 'styled-components'
